requirements.txt
# Word文档处理
python-docx==1.2.0
python-docx-replace==0.1.0

# GUI框架
customtkinter==5.2.0
Pillow==10.0.0

# 配置文件处理
PyYAML==6.0.1
jsonschema==4.20.0

# 日志和工具
loguru==0.7.2
click==8.1.7
tqdm==4.66.1

# macOS 图标支持 (可选)
pyobjc-framework-Cocoa>=12.0; sys_platform == "darwin"
